Moved environments and changed from long values to long longs, now works

+++ day9.cpp
@@ -5,10 +5,10 @@
 
 using namespace std;
 
-long part1(vector<long> input){
-    long invalid = -1;
+long long part1(vector<long long> input){
+    long long invalid = -1;
     for(auto cur = input.begin()+25; cur != input.end(); cur++){
-        long cur_l = *cur;
+        long long cur_l = *cur;
         bool found = false;
         for(auto outer = cur-25; outer != cur && !found; outer++){
             for(auto inner = outer+1; inner != cur && !found; inner++){
@@ -23,12 +23,12 @@ long part1(vector<long> input){
     return invalid;
 }
 
-long part2(vector<long> input){
-    long smallest=0, largest=0;
-    long target = part1(input);
+long long part2(vector<long long> input){
+    long long smallest=0, largest=0;
+    long long target = part1(input);
     for(int i = 2; i < input.size(); i++){
         for(auto cur = input.begin(); cur != input.end(); cur++){
-            long loc_total = 0, loc_sm = *cur, loc_lg = *cur; 
+            long long loc_total = 0, loc_sm = *cur, loc_lg = *cur;
             for(int j = 0; j < i; j++){
                 loc_total += *(cur+j);
                 if(loc_sm > *(cur+j)){
@@ -53,11 +53,11 @@ long part2(vector<long> input){
 
 int main(){
     ifstream input_file("inputs/d9.txt");
-    vector<long> input;
+    vector<long long> input;
     string curr;
     int i = 0;
     while(getline(input_file, curr)){
-        long cur_l = stol(curr);
+        long long cur_l = stoll(curr);
         input.push_back(cur_l);
     }
     cout << part1(input) << endl << part2(input) << endl;
